Refactor:  Separate user and auth logic into distinct microservices
- Created `users` microservice for handling user-related operations.
- Previously, user and authentication logic were combined in the `auth` microservice.
- Separated concerns by moving user management to a dedicated service.
