+++ backend/src/main/java/com/TesisUrbe/backend/Users/Controllers/UserController.java
@@ -1,0 +1,48 @@
+package com.TesisUrbe.backend.Users.Controllers;
+
+import com.TesisUrbe.backend.Users.Services.UserService;
+import com.TesisUrbe.backend.Users.dto.NewUserDto;
+import com.TesisUrbe.backend.Users.exceptions.UserAlreadyExistsException;
+import com.TesisUrbe.backend.security.controllers.AuthController;
+import jakarta.validation.Valid;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.authentication.AuthenticationCredentialsNotFoundException;
+import org.springframework.security.core.Authentication;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.HashMap;
+import java.util.Map;
+
+@RestController
+@RequestMapping("/users")
+public class UserController {
+
+    private final UserService userService;
+    private final AuthController authController;
+
+    public UserController(UserService userService, AuthController authController) {
+        this.userService = userService;
+        this.authController = authController;
+    }
+
+    @PostMapping("/register")
+    public ResponseEntity<Map<String, String>> register(@Valid @RequestBody NewUserDto newUserDto, BindingResult bindingResult, Authentication authentication) {
+        if (bindingResult.hasErrors()) {
+            return ResponseEntity.badRequest().body(authController.errorMap(bindingResult));
+        }
+        try {
+            userService.registerUser(newUserDto, authentication);
+            return ResponseEntity.status(HttpStatus.CREATED).body(Map.of("message", "Usuario registrado con éxito"));
+        } catch (UserAlreadyExistsException e) {
+            return ResponseEntity.badRequest().body(Map.of("error", e.getMessage()));
+        } catch (org.springframework.security.access.AccessDeniedException e) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(Map.of("error", e.getMessage()));
+        } catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(Map.of("error", "Error interno: " + e.getMessage()));
+        }
+    }
+
+}
