+++ backend/src/main/java/com/TesisUrbe/backend/Users/enums/RoleList.java
@@ -1,4 +1,4 @@
-package com.TesisUrbe.backend.security.enums;
+package com.TesisUrbe.backend.Users.enums;
 
 public enum RoleList {
     ROLE_USER, ROLE_ADMIN
