+++ backend/src/main/java/com/TesisUrbe/backend/Users/exceptions/UserAlreadyExistsException.java
@@ -1,4 +1,4 @@
-package com.TesisUrbe.backend.security.exceptions;
+package com.TesisUrbe.backend.Users.exceptions;
 
 public class UserAlreadyExistsException extends RuntimeException {
     public UserAlreadyExistsException(String message) {
