+++ backend/src/main/java/com/TesisUrbe/backend/Users/repository/RoleRepository.java
@@ -1,7 +1,7 @@
-package com.TesisUrbe.backend.security.repository;
+package com.TesisUrbe.backend.Users.repository;
 
-import com.TesisUrbe.backend.security.enums.RoleList;
-import com.TesisUrbe.backend.security.model.Role;
+import com.TesisUrbe.backend.Users.enums.RoleList;
+import com.TesisUrbe.backend.Users.model.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
