+++ backend/src/main/java/com/TesisUrbe/backend/Users/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.TesisUrbe.backend.security.repository;
+package com.TesisUrbe.backend.Users.repository;
 
-import com.TesisUrbe.backend.security.model.User;
+import com.TesisUrbe.backend.Users.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
