--- backend/src/main/java/com/TesisUrbe/backend/security/services/UserService.java
@@ -1,46 +1,0 @@
-package com.TesisUrbe.backend.security.services;
-
-import com.TesisUrbe.backend.security.model.User;
-import com.TesisUrbe.backend.security.repository.UserRepository;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.stereotype.Service;
-
-import java.util.Collections;
-
-@Service
-public class UserService implements UserDetailsService {
-
-    private final UserRepository userRepository;
-
-    public UserService(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-
-    @Override
-    public UserDetails loadUserByUsername(String userName) throws UsernameNotFoundException {
-        User user = userRepository.findByUserName(userName)
-                .orElseThrow(() -> new UsernameNotFoundException("Usuario no encontrado"));
-        SimpleGrantedAuthority authority = new SimpleGrantedAuthority(user.getRole().getName().name());
-
-        return new org.springframework.security.core.userdetails.User(
-                user.getUserName(),
-                user.getPassword(),
-                Collections.singletonList(authority)
-        );
-    }
-
-    public boolean existByUserName(String userName) {
-        return userRepository.existsByUserName(userName);
-    }
-
-    public boolean existByEmail(String email) {
-        return userRepository.existsByEmail(email);
-    }
-
-    public void save(User user) {
-        userRepository.save(user);
-    }
-}
